pyproject.toml
[tool.poetry]
name = "lower-bounded-proper-loss"
version = "0.1.0"
description = ""
authors = ["Shuhei M. Yoshida <s_m_yoshida@nec.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.7.0"
torchvision = "^0.8.1"
pytorch-ignite = "^0.4.2"
omegaconf = "^2.0.5"
tqdm = "^4.54.0"
tensorboardX = "^2.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.790"
isort = "^5.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"